Update direction.py
Added a bunch of documentation, still need to do some more

added a parameter for cycle size, which allows for more variation and experimentation. Avoids hard coding this into the class

changed waiting times to be a ndarray from the start, makes things easy. Made changes to reflect this

added a len method

added setter and getter methods for cycle_size